# backend/docs/swagger.yaml
basePath: /
definitions:
  internal_httpserver_handlers.ErrorResponse:
    properties:
      error:
        type: string
      ok:
        type: boolean
    type: object
  internal_httpserver_handlers.GenericObject:
    additionalProperties: {}
    type: object
...
info:
  contact: {}
  description: 后端基于 Gin 的 HTTP API 文档（swaggo/swag 自动生成）。
  title: F1InkDashboard Backend API
  version: 1.0.0
paths:
...
  /api/v1/f1/sessions:
    get:
      parameters:
      - default: Asia/Shanghai
        description: IANA 时区名称
        in: query
        name: tz
        type: string
      - default: 2026
        description: 赛季年份
        in: query
        name: season
        type: integer
      - description: 分站 round（1-30）
        in: query
        name: round
        type: integer
      - default: auto
        description: session 名称过滤；auto 表示按当前时间选择
        in: query
        name: session
        type: string
      - description: 排位分段（1-3）
        in: query
        name: q
        type: integer
      - default: 13
        description: 返回数量限制（1-30）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 查询 session 列表
      tags:
      - Sessions
  /api/v1/f1/sessions/{season}/{round}/{session_name}:
    get:
      description: 读取静态 session 文件（session_name 会自动去掉 .json 后缀）。
      parameters:
      - description: 赛季年份
        in: path
        name: season
        required: true
        type: integer
      - description: 分站 round（1-30）
        in: path
        name: round
        required: true
        type: integer
      - description: session 名称（可带 .json）
        in: path
        name: session_name
        required: true
        type: string
      - default: Asia/Shanghai
        description: IANA 时区名称
        in: query
        name: tz
        type: string
      - description: 排位分段（1-3）
        in: query
        name: q
        type: integer
      - default: 13
        description: 返回数量限制（1-30）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 按路径获取 session 文件
      tags:
      - Sessions
...
  /api/v1/mp/session-results:
    get:
      description: 返回该 session 排名 + 最快圈等（会拼装车手/车队信息）。
      parameters:
      - description: OpenF1 session_key
        in: query
        name: session_key
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: Session 成绩
      tags:
      - MiniProgram
  /api/v1/mp/telemetry/controls:
    get:
      description: 返回 points[{t, throttle, brake}]。
      parameters:
      - description: OpenF1 session_key
        in: query
        name: session_key
        required: true
        type: integer
      - description: 车手号码
        in: query
        name: driver_number
        required: true
        type: integer
      - default: 320
        description: 采样点数（最大 900）
        in: query
        name: "n"
        type: integer
      - default: fastest
        description: 选择圈：fastest/all 等
        in: query
        name: lap
        type: string
      - description: 指定圈号
        in: query
        name: lap_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 油门/刹车控制曲线
      tags:
      - MiniProgram
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Authorization: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  TokenQuery:
    description: 简单鉴权 token（NEWS_INGEST_TOKEN / OPENF1_INGEST_TOKEN / WECHATPAY_API_TOKEN
      等）
    in: query
    name: token
    type: apiKey
swagger: "2.0"